Added permissions safety

+++ api/childSheet/delete.php
@@ -2,17 +2,27 @@
 
 require dirname( __FILE__ )."/../../config.php";
 
-$removeChildSheetInstance = new ChildSheetController();
-$removeChildSheetInstance->setId($_GET['id']);
-$removeChildSheetInstance->delete();
+if($_SESSION["permission"] == "admin"){
 
-$response = array("status" => 1, "msg" => "Child sheet instance has been deleted." );
+	$removeChildSheetInstance = new ChildSheetController();
+	$removeChildSheetInstance->setId($_GET['id']);
+	$removeChildSheetInstance->delete();
 
-// set response code - 200 OK
-http_response_code(200);
- 
-// show response data in json format
-echo json_encode($response);
+	$response = array("status" => 1, "msg" => "Child sheet instance has been deleted." );
 
+	// set response code - 200 OK
+	http_response_code(200);
+	 
+	// show response data in json format
+	echo json_encode($response);
+
+} else {
+	$response = array("status" => 0, "msg" => "You have no permissions for this operation.");
+	// set response code - 200 OK
+	http_response_code(200);
+	 
+	// show response data in json format
+	echo json_encode($response);
+}
 
 ?>
