+++ api/user/delete.php
@@ -2,17 +2,28 @@
 
 require dirname( __FILE__ )."/../../config.php";
 
-$removeUser = new UserController();
-$removeUser->setId($_GET['id']);
-$removeUser->delete();
+if($_SESSION["permission"] == "admin"){
 
-$response = array("status" => 1, "msg" => "User has been deleted." );
+	$removeUser = new UserController();
+	$removeUser->setId($_GET['id']);
+	$removeUser->delete();
 
-// set response code - 200 OK
-http_response_code(200);
- 
-// show response data in json format
-echo json_encode($response);
+	$response = array("status" => 1, "msg" => "User has been deleted." );
+
+	// set response code - 200 OK
+	http_response_code(200);
+	 
+	// show response data in json format
+	echo json_encode($response);
+
+} else {
+	$response = array("status" => 0, "msg" => "You have no permissions for this operation.");
+	// set response code - 200 OK
+	http_response_code(200);
+	 
+	// show response data in json format
+	echo json_encode($response);
+}
 
 
 ?>
